Safari-609.1.11.1/LayoutTests/editing/images/redo-insert-editable-image-maintains-strokes-expected.txt


Had 0 strokes in editable image before drawing.
Had 1 stroke in editable image after drawing.
Had 0 strokes in editable image after undo.
Had 1 stroke in editable image after redo.
Had 2 strokes in editable image after drawing again.
